added context for making info global
